Extract functions into separate consumables files
